+++ internal/commit/apply.go
@@ -82,6 +82,7 @@ func RunApply(args []string) {
 
 	// run safety checks w/ pre-apply prompt
 	issues := config.RunStaticChecks(cfg)
+	issues = append(issues, config.RunDynamicChecks(cfg)...)
 	errCount := config.PrintIssues(issues)
 
 	if errCount > 0 {