+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.4.1] - 2026-Apr-30
-- 
+- `safety.go` init
+- static firewall logic validations
 
 ## [0.4.0] - 2026-Apr-28
 - `default_rules` adjustments for v6 NDP and DHCPv4/6